requirements.txt
numpy>=1.23
pillow>=10.1.0
torch>=2.1.0
torchvision>=0.18.0
tqdm>=4.64.0